Change fetching policy to be based on AWS default
This commit modifies the policies and roles generated in the fetching
template to be based on the "ReadOnlyAccess" policy that is available in
AWS by default. Two policies are created on top of that: one that denys
access to various actions which are still sensitive, and another which
further allows specific actions that are required.